# specs/001-gw2-resource-farm-bot/contracts/control-api.openapi.yaml
openapi: 3.0.3
info:
  title: GW2 Bot Control API
  version: 0.1.0
  description: Local control API for route farming orchestration and telemetry.
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /v1/run/pause:
    post:
      summary: Pause active run
      operationId: pauseRun
      responses:
        '200':
          description: Run paused
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStateResponse'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/run/resume:
    post:
      summary: Resume paused run
      operationId: resumeRun
      responses:
        '200':
          description: Run resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunStateResponse'
        '409':
          description: Invalid state transition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    RunStatus:
      type: string
      enum: [idle, running, paused, stopping, error]
...
    RunStateResponse:
      type: object
      properties:
        cycle_id:
          type: string
          nullable: true
        route_id:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/RunStatus'
        current_waypoint_index:
          type: integer
          minimum: 0
        started_at_utc:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
      required:
        - status
        - current_waypoint_index
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        remediation:
          type: string
      required: [code, message]